Small corrections to if_indextoname resolution on win32, compiles clean. Backports: r1855864
git-svn-id: https://svn.apache.org/repos/asf/apr/apr/branches/1.7.x@1855865 13f79535-47bb-0310-9956-ffa450edef68

+++ network_io/unix/sockaddr.c
@@ -30,7 +30,7 @@
 #endif
 
 #if defined(HAVE_IF_INDEXTONAME) && defined(_MSC_VER)
-#include <Iphlpapi.h>
+#include "arch/win32/apr_arch_misc.h"
 #endif
 
 #define APR_WANT_STRFUNC
@@ -1252,7 +1252,7 @@ APR_DECLARE(apr_status_t) apr_sockaddr_zone_set(apr_sockaddr_t *sa,
             return APR_EGENERAL;
         }
 
-        sa->sa.sin6.sin6_scope_id = i;
+        sa->sa.sin6.sin6_scope_id = (unsigned int) i;
         return APR_SUCCESS;
     }
 #endif
